[AD-254] Implement FLOOR(... TO WEEK). Split query/DocumentDbQueryMappingServiceTest.java https://bitquill.atlassian.net/browse/AD-254
